backend/pyproject.toml
[project]
name = "efilo-backend"
version = "0.1.0"
description = "efilo.ai — FastAPI backend for compliance-first construction intelligence"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",
    # Auth
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "email-validator>=2.0.0",
    # AI
    "anthropic>=0.40.0",
    "openai>=1.58.0",
    # File storage (R2 / S3)
    "boto3>=1.35.0",
    # Email
    "aiosmtplib>=3.0.0",
    # Background jobs
    "celery[redis]>=5.4.0",
    # Document processing (direct imports!)
    "docling>=2.0.0",
    "chonkie[semantic]>=1.0.0",
    # PDF fallback
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    # Image processing
    "pillow>=11.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "python-cuid2>=1.1.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
